% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{burnInRows}
\alias{burnInRows}
\title{burnInRows}
\usage{
burnInRows(Nthin, Nburn, samples)
}
\arguments{
\item{Nthin}{number of samples to skip before keeping one MCMC sample}

\item{Nburn}{number of samples to discard at the beginning of the chain}

\item{samples}{a 3-d array containing MCMC samples}
}
\value{
3-d array with altered dimension, leaving out Nburn samples and keeping one out of every Nthin
}
\description{
A function that applies row-wise burn-in to the MCMC samples
}
