% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx.R
\name{createFormula}
\alias{createFormula}
\title{createFormula}
\usage{
createFormula(random = NULL, fixed = NULL, intercept = T)
}
\arguments{
\item{random}{a vector containing the names of variables to be treated as random effects}

\item{fixed}{a vector containing the names of variables to be treated as fixed effects}

\item{intercept}{a logical value stating whether the formula should contain an intercept or not, the default is \code{intercept = TRUE}}
}
\value{
a formula object
}
\description{
A function only used inside the \code{remx} function to create a formula that will be passed to \code{glmer} or \code{stan_glmer} functions. It should never be called outside \code{remx}
}
