% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx.R
\name{fixef_metarem}
\alias{fixef_metarem}
\title{fixef_metarem}
\usage{
fixef_metarem(object)
}
\arguments{
\item{object}{a metarem object}
}
\value{
a list containing the posterior means of the fixed effects and their standard deviations
}
\description{
A function to extract fixed effects of a \code{metarem} object. Similar to the function available in \code{lme4} and \code{rstanarm}.
}
