% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx.R
\name{hrem}
\alias{hrem}
\title{hrem}
\usage{
hrem(
  edgelist,
  statistics,
  fixed = NULL,
  random = NULL,
  Niter = 10000,
  Nchain = 2,
  NburnIn = 5000,
  Nthin = 10,
  actors = NULL,
  ordinal = FALSE,
  directed = TRUE
)
}
\arguments{
\item{edgelist}{a list containing multiple relational event sequences}

\item{statistics}{a list containing multiple \code{remstats} objects}

\item{fixed}{a vector containing the names of the covariates that will be treated as fixed effects.}

\item{random}{a vector containing the names of the covariates that will be treated as random effects.}

\item{Niter}{number of iterations of the Gibbs sampler. The default is \code{Niter = 100000}.}

\item{Nchain}{number of chains of the Gibbs sampler.The default is \code{Nchain = 2}.}

\item{NburnIn}{number of samples to be discarded.The default is \code{NburnIn = 5000}.}

\item{Nthin}{number of samples to discarded in between every sample kept.The default is \code{NburnIn = 10}.}

\item{actors}{a list containing the names of the actors in every network, should be the same lenght as \code{edgelist} and \code{statistics}.}

\item{ordinal}{logical value indicating whether the timestamps of the events are know. The default is \code{ordinal = FALSE}. If \code{ordinal = TRUE}, that means that only the order of the events is known.}

\item{directed}{logical value indicating whether the model is for directed network. The default is \code{directed = TRUE}.}
}
\value{
A list containing the output of the Gibbs sampler for this model.
}
\description{
A function to fit tie-oriented mixed-effect relational event models using a meta-analytix approximation. This function is called inside the \code{remx} function, so the user should avoid calling this function and using \code{remx} instead.
}
