% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx.R
\name{hremActor}
\alias{hremActor}
\title{hremActor}
\usage{
hremActor(
  reh,
  statistics,
  Niter = 10000,
  Nchain = 2,
  NburnIn = 5000,
  Nthin = 10,
  random_sender = NULL,
  random_receiver = NULL
)
}
\arguments{
\item{reh}{a list containing multiple relational event sequences pre-processed by the \code{remify} function.}

\item{statistics}{a list containing multiple \code{remstats} objects}

\item{Niter}{number of iterations of the Gibbs sampler. The default is \code{Niter = 100000}.}

\item{Nchain}{number of chains of the Gibbs sampler.The default is \code{Nchain = 2}.}

\item{NburnIn}{number of samples to be discarded.The default is \code{NburnIn = 5000}.}

\item{Nthin}{number of samples to discarded in between every sample kept.The default is \code{NburnIn = 10}.}

\item{random_sender}{a vector containing the names of the covariates that will be treated as random effects in the sender model.}

\item{random_receiver}{a vector containing the names of the covariates that will be treated as random effects in the receiver model.}
}
\value{
A list containing the output of the Gibbs sampler for this model.
}
\description{
A function to fit DyNaM mixed-effect relational event models using a meta-analytic approximation. This function is called inside the \code{remx} function, so the user should avoid calling this function and using \code{remx} instead.
}
