% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx.R
\name{intervals}
\alias{intervals}
\title{interval}
\usage{
intervals(fit, prob = 0.95)
}
\arguments{
\item{fit}{A streamrem object.}

\item{prob}{A probability to compute the confidence interval, default is 0.95.}
}
\description{
interval
}
