% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat2DF}
\alias{mat2DF}
\title{mat2DF}
\usage{
mat2DF(x)
}
\arguments{
\item{x}{a matrix to be converted into a data frame}
}
\value{
x converted to a data frame
}
\description{
A function that converts a matrix to data frame
}
