% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{networks}
\alias{networks}
\title{networks}
\format{
A list containing 15 data frames, each data frame contains:
\describe{
\item{time}{the time at which each observation occurred}
\item{sender}{the index representing the sender of the interaction}
\item{receiver}{the index represending the receiver of the interaction}
}
}
\usage{
networks
}
\description{
A set of 15 randomly generated networks. See https://github.com/TilburgNetworkGroup/remulate.
}
\keyword{datasets}
