% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx.R
\name{plot.streamrem}
\alias{plot.streamrem}
\title{plot.streamrem}
\usage{
\method{plot}{streamrem}(x, same_page = FALSE, ...)
}
\arguments{
\item{x}{A streamrem object.}

\item{same_page}{The default is FALSE. If TRUE, the function will create all graphs in one plot.}

\item{...}{further arguments to be passed to the 'summary' method}
}
\description{
Plots lines with the trends of each parameter estimates and 95% confidence intervals.
}
