% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx.R
\name{print.streamrem}
\alias{print.streamrem}
\title{print.streamrem}
\usage{
\method{print}{streamrem}(x, ...)
}
\arguments{
\item{x}{is a \code{summary.streamrem} object}

\item{...}{further arguments to be passed to the 'print.summary' method}
}
\description{
A function that prints out a summary of a 'streamrem' object.
}
\details{
print.streamrem
}
