% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx.R
\name{print.summary.metarem}
\alias{print.summary.metarem}
\title{print.summary.metarem}
\usage{
\method{print}{summary.metarem}(x, ...)
}
\arguments{
\item{x}{is a \code{summary.metarem} object}

\item{...}{further arguments to be passed to the 'print.summary' method}
}
\description{
A function that prints out a summary of a 'metarem' object.
}
\details{
print.summary.metarem
}
