% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx.R
\name{ranef_metarem}
\alias{ranef_metarem}
\title{ranef_metarem}
\usage{
ranef_metarem(object)
}
\arguments{
\item{object}{a metarem object}
}
\value{
a list containing the posterior means of the random effects and their standard deviations
}
\description{
A function to extract random effects of a \code{metarem} object. Similar to the function available in \code{lme4} and \code{rstanarm}.
}
