% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rehPoisson}
\alias{rehPoisson}
\title{rehPoisson}
\usage{
rehPoisson(stats)
}
\arguments{
\item{stats}{a remstats object, the objects inside the list must be in the same order as they would be in remstats}
}
\value{
a data frame with events stacked
}
\description{
A function that reshapes the data set to allow implementation of the REM writen as Poisson regression
}
