% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx-package.R
\docType{package}
\name{remx-package}
\alias{remx-package}
\alias{_PACKAGE}
\title{remx: Mixed-effect relational event models}
\description{
This package implements meta-analytic approximations for mixed-effect tie- and actor-oriented relational event models. Also, using a fully frequentist model will return and lme4 object and using fully Bayesian model will return an rstanarm object.
}
\author{
\strong{Maintainer}: Fabio Generoso Vieira \email{f.v.generosovieira@tilburguniversity.edu}

Authors:
\itemize{
  \item Rumana Lakdawala \email{r.j.lakdawala@tilburguniversity.edu}
  \item Marlyne Meijerink-Bosman \email{m.l.meijerink@tilburguniversity.edu}
  \item Giuseppe Arena \email{g.arena@tilburguniversity.edu}
  \item Diana Karimova \email{d.karimova@tilburguniversity.edu}
  \item Mahdi Shafiee Kamalabad \email{m.shafiee@tilburguniversity.edu}
}

Other contributors:
\itemize{
  \item Roger Leenders \email{r.t.a.j.leenders@tilburguniversity.edu} [contributor]
  \item Joris Mulder \email{j.mulder3@tilburguniversity.edu} [contributor]
}

}
\keyword{internal}
