% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rndMeans}
\alias{rndMeans}
\title{rndMeans
This function only returns the components of the fixed effects that are random effects means}
\usage{
rndMeans(x, psi)
}
\arguments{
\item{x}{matrix specifying which parameters are to be extracted from the vector}

\item{psi}{a vector containing fixed- and random-effects estimates}
}
\description{
rndMeans
This function only returns the components of the fixed effects that are random effects means
}
