% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampler}
\alias{sampler}
\title{sampler}
\usage{
sampler(
  Niter,
  Nchain,
  NburnIn,
  Nthin,
  p,
  q,
  K,
  delta,
  beta_hat,
  omega_hat,
  omega_hat_11,
  omega_hat_21,
  omega_hat_22,
  Mu,
  Sigma,
  eta,
  Lambda,
  xi,
  SigmaMean,
  psi,
  psi_hat,
  random_effect,
  fixed_effect,
  randomModel
)
}
\arguments{
\item{Niter}{number of iterations to run the Gibbs sampler}

\item{Nchain}{number of MCMC chains}

\item{NburnIn}{number of samples to discard at the beginning of every chain}

\item{Nthin}{number of samples to skip for every MCMC sample kept}

\item{p}{number of random effects in the model}

\item{q}{number of fixed effects in the model}

\item{K}{number of groups (or networks) in the multilevel analysis}

\item{delta}{a 3-d array with dimension p x K x Nchain containing the starting values for the random effects parameters}

\item{beta_hat}{a matrix with dimensions p X K containing the MLEs of the random effects estimated from the independent networks}

\item{omega_hat}{a 3-d array with dimensions p x p x K containing the covariance matrices of the MLEs estimated from the independent groups}

\item{omega_hat_11}{covariance matrix decomposition of the fixed effects}

\item{omega_hat_21}{covariance matrix decomposition of the covariance between fixed and random effects}

\item{omega_hat_22}{covariance matrix decomposition of the random effects}

\item{Mu}{a matrix with dimensions Nchain x p containing the initial values for the random-effect means}

\item{Sigma}{a 3-d array with dimensions p x p x Nchain containing the initial values for the random-effect covariance matrix}

\item{eta}{an integer that we set equal to 2, so the covariances have an uniform prior, see Huang & Wand (2013)}

\item{Lambda}{a matrix with dimensions Nchain x p containing initial values for the Lambda parameter}

\item{xi}{a scalar that can be an arbitrarily large number, see Gelman (2006)}

\item{SigmaMean}{a matrix with dimensions q x K representing the standard errors of the MLEs for the fixed effects}

\item{psi}{a matrix with dimensions q x K representing the MLEs for the fixed effects independently estimated from every group}

\item{psi_hat}{a matrix with dimensions q x K representing the MLEs for the fixed effects independently estimated from every group}

\item{random_effect}{a matrix with 1's and 0's}

\item{fixed_effect}{a matrix with 1's and 0's}

\item{randomModel}{a boolean, if equals to TRUE it states that the model contain only random effects}
}
\value{
a list containing the MCMC samples for all parameters in the model
}
\description{
A function that runs the Gibbs sampler for a tie-oriented mixed-effect relational event model
}
