% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stream}
\alias{stream}
\title{stream}
\format{
A data frame with 3 columns for time, sender and receiver
\describe{
\item{time}{the time at which each observation occurred}
\item{actor1}{the index representing the sender of the interaction}
\item{actor2}{the index represending the receiver of the interaction}
}
}
\usage{
stream
}
\description{
A network with 5000 events. See https://github.com/TilburgNetworkGroup/remulate.
}
\keyword{datasets}
