% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx.R
\name{summary.metarem}
\alias{summary.metarem}
\title{summary.metarem}
\usage{
\method{summary}{metarem}(object, ...)
}
\arguments{
\item{object}{is a \code{metarem} object}

\item{...}{further arguments to be passed to the 'summary' method}
}
\description{
A function that arranges a summary of a 'metarem' object
}
\details{
summary.metarem
}
