% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx.R
\name{summary.streamrem}
\alias{summary.streamrem}
\title{summary.streamrem}
\usage{
\method{summary}{streamrem}(object, ...)
}
\arguments{
\item{object}{is a \code{streamrem} object}

\item{...}{further arguments to be passed to the 'summary' method}
}
\description{
A function that arranges a summary of a 'streamrem' object
}
\details{
summary.streamrem
}
