% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdateDelta}
\alias{UpdateDelta}
\title{UpdateDelta
This is the function that samples the random effects}
\usage{
UpdateDelta(
  q,
  p,
  K,
  beta_hat,
  Mu,
  omega_hat_11,
  omega_hat_21,
  omega_hat_22,
  psi_hat,
  psi,
  Sigma
)
}
\arguments{
\item{q}{number of fixed effects}

\item{p}{number of random effects}

\item{K}{number of networks}

\item{beta_hat}{MLE estimates}

\item{Mu}{mean of the random effects}

\item{omega_hat_11}{covariance matrix decomposition of the fixed effects}

\item{omega_hat_21}{covariance matrix decomposition of the covariance between fixed and random effects}

\item{omega_hat_22}{covariance matrix decomposition of the random effects}

\item{psi_hat}{fixed effects}

\item{psi}{initial values or previous MCMC sample of fixed effects}

\item{Sigma}{covariance matrix of the random effects}
}
\description{
UpdateDelta
This is the function that samples the random effects
}
