% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdateDeltaInd}
\alias{UpdateDeltaInd}
\title{UpdateDeltaInd}
\usage{
UpdateDeltaInd(p, K, beta_hat, omega_hat_22, Mu, Sigma)
}
\arguments{
\item{p}{number of random effects}

\item{K}{number of networks}

\item{beta_hat}{MLE estimates}

\item{omega_hat_22}{covariance matrix of the random effects}

\item{Mu}{random effects mean}

\item{Sigma}{covariance matrix of the random effects}
}
\description{
UpdateDeltaInd
}
