% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdateLambda}
\alias{UpdateLambda}
\title{UpdateLambda}
\usage{
UpdateLambda(p, eta, Sigma, xi)
}
\arguments{
\item{p}{an integer denoting the dimension of the covariance matrix Sigma}

\item{eta}{an integer that we set equal to 2, so the covariances have an uniform prior, see Huang & Wand (2013)}

\item{Sigma}{a p x p matrix representing one MCMC sample of the random-effect covariance matrix}

\item{xi}{a scalar that can be an arbitrarily large number, see Gelman (2006)}
}
\value{
a row vector representing one MCMC sample of the mixture parameter that allows the covariance matrix of the random effects to have a Half-t prior
}
\description{
A function to update the mixture parameters that allow Sigma to have a Halt-t prior, see Huang & Wand (2013)
}
