% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdateMean}
\alias{UpdateMean}
\title{UpdateMean
This is a new function that models fixed-effects and random-effects means together so we do not ignore their dependency
This function already does what we want}
\usage{
UpdateMean(q, p, K, SigmaMean, psi, delta, randomModel)
}
\arguments{
\item{q}{number of fixed effects}

\item{p}{number of random effects}

\item{K}{number of networks}

\item{SigmaMean}{covariance matrices of the MLE estimates}

\item{psi}{MLE estimates}

\item{delta}{random-effects estimates}

\item{randomModel}{boolean, if TRUE model contains random effects}
}
\description{
UpdateMean
This is a new function that models fixed-effects and random-effects means together so we do not ignore their dependency
This function already does what we want
}
