% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdateOmega}
\alias{UpdateOmega}
\title{UpdateOmega}
\usage{
UpdateOmega(p, K, beta_hat, omega_hat, Mu, Sigma, nu, beta)
}
\arguments{
\item{p}{an integer that determines the dimension of the vector of parameters in each group}

\item{K}{an integer representing the number of groups in the analysis}

\item{beta_hat}{a matrix with dimensions p x K representing the MLE estimates for each independent groups}

\item{omega_hat}{a 3-d array with dimensions p x p x K representing the covariance matrix of the MLEs}

\item{Mu}{a vector with dimension p representing the random-effect means}

\item{Sigma}{a matrix p x p representing the covariance matrix of the random effects}

\item{nu}{a vector of dimension K representing the degress of freedom of the student-t distribution for each group}

\item{beta}{a matrix with dimensions p x K representing one MCMC sample of the group-specific parameters, aka random effects}
}
\value{
a vector with dimension K representing one MCMC sample of the scale mixture variable omega
}
\description{
A function to update the scale mixture variable that allows beta to have a student-t likelihood
}
