% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdatePsi}
\alias{UpdatePsi}
\title{UpdatePsi}
\usage{
UpdatePsi(q, K, TauPsi, SigmaPsi, psi)
}
\arguments{
\item{q}{an integer representing the number of fixed effects in the model}

\item{K}{an integer representing the number of groups (or networks) in the multilevel analysis}

\item{TauPsi}{a vector of dimension q, which represents the prior variance of the fixed effects}

\item{SigmaPsi}{a matrix with dimensions q x K representing the standard errors of the MLEs for the fixed effects}

\item{psi}{a matrix with dimensions q x K representing the MLEs for the fixed effects independently estimated from every group}
}
\value{
a row vector with dimension q representing one MCMC sample of the fixed effects
}
\description{
A function that updates the fixed-effect parameters in the mixed-effect relational event model
}
