% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdateSigma}
\alias{UpdateSigma}
\title{UpdateSigma}
\usage{
UpdateSigma(eta, K, p, Lambda, beta, Mu)
}
\arguments{
\item{eta}{an integer representing a prior hyperparameter, this will allow Sigma to have a Half-t prior}

\item{K}{an integer representing the number of groups}

\item{p}{an integer representing the number of the covariates in the models, Sigma in the end will have dimensions p x p}

\item{Lambda}{a vector with dimension p, this is an auxiliary variable that allows Sigma to have a Half-t prior}

\item{beta}{is a matrix with dimensions p x K representing one MCMC sample of the random effects}

\item{Mu}{is a vector with dimension p representing the mean of the random effects}
}
\value{
a p x p matrix representing one MCMC sample of the covariance matrix of the random effects
}
\description{
A function to update the covariance matrix of the random effects, for details see Huang & Wand (2013)
}
