% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx.R
\name{VarCov_metarem}
\alias{VarCov_metarem}
\title{VarCov_metarem}
\usage{
VarCov_metarem(object)
}
\arguments{
\item{object}{a metarem object}
}
\value{
a matrix of the posterior means of the covariance matrix of the random effects.
}
\description{
A function to extract the covariance matrix of the random effects of a \code{metarem} object.
}
