% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx.R
\name{print.strem}
\alias{print.strem}
\title{print.strem}
\usage{
\method{print}{strem}(x, ...)
}
\arguments{
\item{x}{is a \code{summary.strem} object}

\item{...}{further arguments to be passed to the 'print.summary' method}
}
\description{
A function that prints out a summary of a 'strem' object.
}
\details{
print.strem
}
