% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx.R

\name{print.summary.strem}
\alias{print.summary.strem}
\title{print.summary.strem}
\usage{
\method{print}{summary.strem}(x, ...)
}
\arguments{
\item{x}{is a \code{summary.strem} object}


\item{...}{further arguments to be passed to the 'print.summary' method}
}
\description{
A function that prints out a summary of a 'metarem' object.
}
\details{
print.summary.strem
}
