% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{samplerActor}
\alias{samplerActor}
\title{samplerActor}
\usage{
samplerActor(
  Niter,
  Nchain,
  NburnIn,
  Nthin,
  v,
  u,
  K,
  nu,
  beta,
  beta_hat,
  omega_hat,
  MuB,
  Sigma,
  eta,
  LambdaB,
  xi,
  TauMuB,
  TauPsi,
  SigmaPsi,
  omegaB,
  psi,
  p,
  q,
  gamma,
  gamma_hat,
  zeta_hat,
  MuG,
  Zeta,
  LambdaG,
  TauMuG,
  TauPhi,
  SigmaPhi,
  omegaG,
  phi,
  randomModelSnd,
  randomModelRec,
  fixModelSnd,
  fixModelRec
)
}
\arguments{
\item{Niter}{number of iterations to run the Gibbs sampler}

\item{Nchain}{number of MCMC chains}

\item{NburnIn}{number of samples to discard at the beginning of every chain}

\item{Nthin}{number of samples to skip for every MCMC sample kept}

\item{v}{an integer representing the number of random effects in the receiver model}

\item{u}{an integer representing the number of fixed effects in the receiver model}

\item{K}{number of groups (or networks) in the multilevel analysis}

\item{nu}{a vector with dimension K, representing the degrees of freedom of the student-t distribution, it is set to (number of events in every group) - p}

\item{beta}{a 3-d array with dimension v x K x Nchain containing the starting values for the random effects parameters in the receiver model}

\item{beta_hat}{a matrix with dimensions v X K containing the MLEs of the random effects in the receiver model estimated from the independent networks}

\item{omega_hat}{a 3-d array with dimensions v x v x K containing the covariance matrices of the MLEs for the random effects in the receiver model estimated from the independent groups}

\item{MuB}{a matrix with dimensions v x Nchain containing the initial values of the random-effect means in the receiver model}

\item{Sigma}{a 3-d array with dimensions v x v x Nchain containing the initial values for the covariance matrix of the random-effects in the receiver model}

\item{eta}{an integer that we set equal to 2, so the covariances have an uniform prior, see Huang & Wand (2013)}

\item{LambdaB}{a matrix with dimensions Nchain x v containing initial values for the Lambda parameter in the receiver model}

\item{xi}{a scalar that can be an arbitrarily large number, see Gelman (2006)}

\item{TauMuB}{is a matrix with dimensions v x v representing the covariance matrix of the prior of Mu in the receiver model}

\item{TauPsi}{a vector of dimension u, which represents the prior variance of the fixed effects in the receiver model}

\item{SigmaPsi}{a matrix with dimensions u x K representing the standard errors of the MLEs for the fixed effects in the receiver model}

\item{omegaB}{a matrix with dimensions v x Nchain containing the initial values for the scale mixture parameters that allow the random-effects in the receiver model to have a student-t likelihood}

\item{psi}{a matrix with dimensions u x K representing the MLEs for the fixed effects in the receiver model independently estimated from every group}

\item{p}{an integer representing the number of random effects in the sender model}

\item{q}{an integer representing the number of fixed effects in the sender model}

\item{gamma}{a 3-d array with dimension p x K x Nchain containing the starting values for the random effects parameters in the sender}

\item{gamma_hat}{a matrix with dimensions p X K containing the MLEs of the random effects in the sender model estimated from the independent networks}

\item{zeta_hat}{a 3-d array with dimensions p x p x K containing the covariance matrices of the MLEs for the random effects in the sender model estimated from the independent groups}

\item{MuG}{a matrix with dimensions p x Nchain containing the initial values of the random-effect means in the sender model}

\item{Zeta}{a 3-d array with dimensions p x p x Nchain containing the initial values for the covariance matrix of the random-effects in the sender model}

\item{LambdaG}{a matrix with dimensions Nchain x p containing initial values for the Lambda parameter in the sender model}

\item{TauMuG}{is a matrix with dimensions p x p representing the covariance matrix of the prior of Mu in the sender model}

\item{TauPhi}{a vector of dimension q, which represents the prior variance of the fixed effects in the sender model}

\item{SigmaPhi}{a matrix with dimensions q x K representing the standard errors of the MLEs for hte fixed effects in the sender model}

\item{omegaG}{matrix with dimensions p x Nchain containing the initial values for the scale mixture parameters that allow the random-effects in the receiver model to have a student-t likelihood}

\item{phi}{a matrix with dimensions q x K representing the MLEs for the fixed effects in the sender model independently estimated from every group}

\item{randomModelSnd}{a boolean, if equals to TRUE it states that the sender model contain only random effects}

\item{randomModelRec}{a boolean, if equals to TRUE it states that the receiver model contain only random effects}

\item{fixModelSnd}{a boolean, if equals TRUE it states that the sender model contain only fixed effects}

\item{fixModelRec}{a boolean, if equals TRUE it states that the receiver model contain only fixed effects}
}
\value{
a list containing two other lists with the MCMC sample for all parameters in the sender and receiver models
}
\description{
A function that runs the Gibbs sampler for the DyNaM (actor-oriented) mixed-effect relational event model
}
