% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remx.R
\name{summary.strem}
\alias{summary.strem}
\title{summary.strem}
\usage{
\method{summary}{strem}(object, ...)
}
\arguments{
\item{object}{is a \code{strem} object}

\item{...}{further arguments to be passed to the 'summary' method}
}
\description{
A function that arranges a summary of a 'strem' object
}
\details{
summary.strem
}
