% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdateBeta}
\alias{UpdateBeta}
\title{UpdateBeta}
\usage{
UpdateBeta(p, K, beta_hat, omega_hat, Mu, Sigma)
}
\arguments{
\item{p}{an integer that determines the dimension of the vector of parameters in each group}

\item{K}{an integer representing the number of groups in the analysis}

\item{beta_hat}{a matrix with dimensions p x K representing the MLE estimates for each independent groups}

\item{omega_hat}{a 3-d array with dimensions p x p x K representing the covariance matrix of the MLEs}

\item{Mu}{a vector p x 1 representing the mean of the random effects}

\item{Sigma}{a matrix p x p representing the covariance matrix of the random effects}
}
\value{
a matrix with dimensions p x K representing one sample of the posterior distribution of the beta
}
\description{
A function to update the group specific parameters of the multilevel model
}
